@@ -22,42 +22,43 @@
   "author": "Tien Do <tiendq@gmail.com>",
   "license": "MIT",
   "dependencies": {
-    "cuid": "^1.3.8",
-    "dropbox": "^2.5.0",
-    "file-loader": "^0.10.0",
+    "cuid": "^2.1.8",
+    "dropbox": "^4.0.30",
+    "file-loader": "^6.0.0",
     "lodash.throttle": "^4.1.1",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
-    "react-redux": "^4.4.6",
-    "redux": "^3.6.0",
-    "redux-devtools": "^3.3.1",
-    "redux-logger": "^2.7.4"
+    "prop-types": "^15.7.2",
+    "react": "16.13.1",
+    "react-dom": "16.13.1",
+    "react-redux": "^7.2.0",
+    "redux": "^4.0.5",
+    "redux-devtools": "^3.5.0",
+    "redux-logger": "^3.0.6"
   },
   "devDependencies": {
-    "babel-core": "^6.17.0",
-    "babel-eslint": "^7.0.0",
-    "babel-jest": "^16.0.0",
-    "babel-loader": "^6.2.5",
-    "babel-plugin-transform-object-rest-spread": "^6.20.2",
-    "babel-preset-es2015": "^6.16.0",
-    "babel-preset-react": "^6.16.0",
-    "css-loader": "^0.25.0",
-    "enzyme": "^2.6.0",
-    "eslint": "^3.8.1",
-    "eslint-loader": "^1.6.0",
-    "eslint-plugin-import": "^2.0.1",
-    "eslint-plugin-jsx-a11y": "^2.2.3",
-    "eslint-plugin-react": "^6.4.1",
+    "@babel/cli": "^7.8.4",
+    "@babel/core": "^7.9.0",
+    "@babel/preset-env": "^7.9.0",
+    "@babel/preset-react": "^7.9.1",
+    "babel-eslint": "^10.1.0",
+    "babel-jest": "^25.1.0",
+    "babel-loader": "^8.1.0",
+    "css-loader": "3.4.2",
+    "enzyme": "3.11.0",
+    "eslint": "6.8.0",
+    "eslint-loader": "3.0.3",
+    "eslint-plugin-import": "2.20.1",
+    "eslint-plugin-jsx-a11y": "6.2.3",
+    "eslint-plugin-react": "7.19.0",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^16.0.2",
-    "node-sass": "^4.2.0",
-    "react-addons-test-utils": "^15.3.2",
-    "react-dom": "^15.3.2",
-    "react-hot-loader": "^3.0.0-beta.6",
-    "react-test-renderer": "^15.3.2",
-    "sass-loader": "^4.0.2",
-    "style-loader": "^0.13.1",
-    "webpack": "^1.13.2",
-    "webpack-dev-server": "^1.16.2"
+    "jest": "25.1.0",
+    "node-sass": "^4.13.1",
+    "react-addons-test-utils": "15.6.2",
+    "react-hot-loader": "4.12.20",
+    "react-test-renderer": "16.13.1",
+    "sass-loader": "8.0.2",
+    "style-loader": "1.1.3",
+    "webpack": "4.42.0",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.10.3"
   }
 }
