@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { createStore, combineReducers, applyMiddleware, compose } from 'redux';
 import { Provider } from 'react-redux';
-import createLogger from 'redux-logger';
+import { createLogger } from 'redux-logger';
 import throttle from 'lodash.throttle';
 import reducers from './components/reducers';
 import { loadState, saveState } from './components/local-storage';
